Add files via upload
The files above are the simple code I've written for a tool project in unity. The game itself is also very simple, and the goal of the project was to apply some of what I learned about tools programming. I have also attached a video to show how these tools work in the game in Unity.